Authorization of the ads (Edit/Delete).
Only the owner of an ad is able to edit/delete the ad.

+++ server/routes/ads.js
@@ -1,20 +1,10 @@
 const express = require('express');
 const router = express.Router();
 const Ad = require('../models/ad');
-const {adSchema} = require('../schemas.js');
-const {isLoggedIn} = require('../middleware.js');
-const ExpressError = require('../utils/ExpressError');
+// const {adSchema} = require('../schemas.js');
+const {isLoggedIn, validateAd, isAuthor} = require('../middleware.js');
+// const ExpressError = require('../utils/ExpressError');
 const catchAsync = require('../utils/catchAsync');
-
-const validateAd = (req, res, next) => {
-    const {error} = adSchema.validate(req.body);
-    if (error) {
-        const msg = error.details.map(el => el.message).join(',');
-        throw new ExpressError(msg, 400);
-    } else {
-        next();
-    }
-}
 
 router.get('/', isLoggedIn, catchAsync(async (req, res) => {
     const ads = await Ad.find({});
@@ -27,9 +17,9 @@ router.get('/new', isLoggedIn, (req, res) => {
 
 router.post('/', validateAd, isLoggedIn, catchAsync(async(req, res) => {
     const ad = new Ad(req.body.ad);
-    ad.author = req.user._id;
+    ad.author = req.user._id; // Giving the new created ad an owner!
     await ad.save();
-    req.user.ads.push(ad._id);
+    req.user.ads.push(ad._id); // Adding the new created ad to the list of the owner's ads
     await req.user.save();
     req.flash('success', 'Successfully made a new Ad!');
     res.redirect(`/ads/${ad._id}`);
@@ -46,8 +36,8 @@ router.get('/:id', isLoggedIn, catchAsync(async (req, res) => {
     res.render('ads/show', {ad});
 }))
 
-router.get('/:id/edit', isLoggedIn, catchAsync(async(req, res) => {
-    const id = req.params.id;
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync(async(req, res) => {
+    const {id} = req.params;
     const ad = await Ad.findById(id);
     if (!ad) {
         req.flash('error', "Can't find that ad!");
@@ -56,15 +46,22 @@ router.get('/:id/edit', isLoggedIn, catchAsync(async(req, res) => {
     res.render('ads/edit', {ad});
 }))
 
-router.put('/:id', validateAd, isLoggedIn, catchAsync(async(req, res) => {
+router.put('/:id', isLoggedIn, isAuthor, validateAd, catchAsync(async(req, res) => {
     const {id} = req.params;
     const ad = await Ad.findByIdAndUpdate(id, {...req.body.ad}, {runValidators: true, new: true});
     req.flash('success', 'Ad is successfully edited!');
     res.redirect(`/ads/${ad._id}`);
 }))
 
-router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
+router.delete('/:id', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
     const {id} = req.params;
+    // First find the ad:
+    const ad = await Ad.findById(id);
+    // Delete the found ad ONLY IF the logged-in user is the owner of the ad:
+    if (!ad.author.equals(req.user._id)) {
+        req.flash('error', "You don't have the permission!");
+        return res.redirect(`/ads/${id}`);
+    }
     await Ad.findByIdAndDelete(id);
     req.flash('success', 'Ad is successfully deleted!');
     res.redirect('/ads');
